backend/requirements.txt
# FarmerClaw Backend Dependencies
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.18

# Database
sqlalchemy==2.0.36
aiosqlite==0.22.0
asyncpg==0.30.0

# LLM
dashscope==1.20.14
openai==1.58.1

# HTTP
httpx==0.28.1

# Utilities
python-dotenv==1.0.1
loguru==0.7.3
tenacity==9.0.0
